requirements.txt
# Основные зависимости для приложения Strod-Service Technology
PyQt5>=5.15.9
SQLAlchemy>=2.0.4
openpyxl>=3.1.2
python-docx>=0.8.11
requests>=2.28.2
psycopg2-binary>=2.9.5
alembic>=1.10.2
python-dotenv>=0.19.2
Pillow>=9.4.0
pyinstaller>=5.13.0